requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
python-dotenv
web3>=7.0.0
